added comments

+++ include/lead_checks.h
@@ -2,8 +2,9 @@
 
 #include "add_structs.h"
 
-/*This additional class saves average values of variables and
- * calculates average value
+/**
+ * @brief This additional class saves average values
+ * of variables and calculates average value
  */
 
 class average_value
@@ -12,15 +13,22 @@ private:
 	
 	//for analysis used last 3 number
 	const size_t N = 3;
-	double* list_value = new double[N];
-	double sum = 0.;
+	float* list_value = new float[N];
+	float sum = 0.f;
 	int  place = 0;
 	
 
 public:
+
 	average_value() {};
 	virtual ~average_value();
-	void set(double);
-	void set_start_value(double);
-	double get();
+
+	///to add values for calculating
+	void set(float);
+
+	/// set value firstly
+	void set_start_value(float);
+	///get an average value
+	float get();
+
 };
